Remove all `uint()` calls

+++ base/string.jl
@@ -660,7 +660,7 @@ const memhash = Uint === Uint64 ? :memhash_seed : :memhash32_seed
 const memhash_seed = Uint === Uint64 ? 0x71e729fd56419c81 : 0x56419c81
 
 function hash{T<:ByteString}(s::Union(T,SubString{T}), h::Uint)
-    h += uint(memhash_seed)
+    h += memhash_seed
     ccall(memhash, Uint, (Ptr{Uint8}, Csize_t, Uint32), s, sizeof(s), itrunc(Uint32,h)) + h
 end
 hash(s::String, h::Uint) = hash(bytestring(s), h)
